.idea/Ochem_Survival/requirements.txt
pygame
rdkit
pillow
openpyxl
pandas
sqlite3